requirements.txt
# Gmail Job Monitor - Dependencies

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Anthropic Claude API
anthropic>=0.39.0

# Environment and configuration
python-dotenv>=1.0.0

# HTML parsing for email content
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Scheduling (for cron-like functionality)
schedule>=1.2.0
